Fixing "TypeError" to some inputs
When the token was an empty or some random input with less than 39 chars, `bufferpack.unpack` method was returning `undefined` - breaking the `Array.shift` used to get the version, timestamp and nonce - throwing "TypeError: Cannot read property 'shift' of undefined"error.

This commit adds a guard clause to this TypeError and a more descriptive error message.